requirements.txt
black
pytest
pytest-cov
pylint
ruff
requests
pandas
python-dotenv
databricks-sql-connector